+++ resolwe/flow/executors/docker/constants.py
@@ -3,3 +3,6 @@ DATA_VOLUME = '/data'
 DATA_ALL_VOLUME = '/data_all'
 UPLOAD_VOLUME = '/upload'
 SECRETS_VOLUME = '/secrets'
+
+# Container image used if no image is specified.
+DEFAULT_CONTAINER_IMAGE = 'resolwe/base:ubuntu-16.04'